backend/requirements.txt
# Core Framework
fastapi
uvicorn[standard]
python-dotenv
pydantic

# LangChain & LangGraph
langgraph
langchain
langchain-google-genai
langchain-openai
langchain-anthropic
langchain-community
langgraph-checkpoint-sqlite
aiosqlite

# Tools & Utilities
httpx
aiohttp

# Data Processing & Charts
pandas
matplotlib
numpy

# Search API
tavily-python

# Testing
pytest
pytest-asyncio

# WebSocket support (included in uvicorn[standard])
websockets
tabulate

# MCP (Model Context Protocol)
mcp
fastmcp

# SSE support
sse-starlette

# RAG & Vector Store
chromadb
pypdf

# Observability (OpenTelemetry)
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
